# Conflicts:
#	TheAlbianTimes/Assets/Scripts/Countries/CountryEvents/CountryEvent.cs
#	TheAlbianTimes/Assets/Scripts/Managers/CountryEventManager.cs
#	TheAlbianTimes/Assets/Scripts/Managers/GameManager.cs
#	TheAlbianTimes/Assets/Scripts/Managers/TestGameManager.cs

+++ TheAlbianTimes/Assets/Scripts/GameState/GameState.cs
@@ -1,6 +1,6 @@
 public class GameStatePlayerData
 {
-    public int money = 1000;
+    public float money = 1000f;
     public int staff = 10;
     public int reputation = 50;
 }
